companycrawl/companycrawl/requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.7.22
certstream==1.12
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
configparser==5.3.0
constantly==15.1.0
crayons==0.4.0
cryptography==38.0.4
cssselect==1.2.0
dnspython==2.3.0
exceptiongroup==1.0.4
fake-useragent==1.1.1
filelock==3.8.2
gevent==22.10.2
google-auth==2.16.1
greenlet==2.0.1
gspread==3.4.2
gspread-dataframe==3.2.2
h11==0.14.0
h2==3.2.0
hpack==3.0.0
httplib2==0.21.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.1
incremental==22.10.0
iplookup==1.0.5
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
Levenshtein==0.20.8
lxml==4.9.2
numpy==1.21.6
oauth2client==4.1.3
opencv-python==4.6.0.66
outcome==1.2.0
packaging==22.0
pandas==1.3.5
parsel==1.7.0
Pillow==9.3.0
priority==1.3.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.20.8
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
queuelib==1.6.2
rapidfuzz==2.13.6
requests==2.28.1
requests-file==1.5.1
rsa==4.9
scipy==1.1.0
Scrapy==2.5.0
scrapy-splash==0.7.2
selenium==4.7.2
selenium-wire==3.0.6
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
telegram==0.0.1
termcolor==2.1.1
tld==0.12.6
tldextract==3.1.0
tornado==6.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==21.2.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
w3lib==2.1.1
webdriver-manager==3.4.2
websocket==0.2.1
websocket-client==0.58.0
websockets==9.1
wincertstore==0.2
wsproto==1.2.0
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
